FireTaskWrapper refactor
